start_mon: Reworking network detection
The actual code is using a lot of regexp and a complex logic to define if we should use ipv4 or ipv6.

This patch try to simplify the read of the start_mon() function by:
- adding basic functions to extract ip or network information
- embedded the logic if 'ip' is missing directly in those functions
- improve the ipv6 support
- simplify the logic by making common calls between ipv4|6

First, the get_ip & get_network functions are created to externalize
the detection code from the main function. This way the logic is easier
to read.

This logic is simplified by unificating cases where NETWORK_AUTO_DETECT > 1 (aka 4 or 6).
We now have the same code base for both.

The newly created functions embedded the logic to detect the network configuration regarding if 'ip' is present or not.

For IPV6, some code is added to read ipv6_route or if_inet6 from /proc to extract the relevant information.
A flat_to_ipv6 function is added to transform ipv6 address in a short expression.

This patch embedded a fix for the NETWORK_AUTO_DETECT=6 case where the network was detecting by using an IPv4 regexp.